+++ airflow/notifications/basenotifier.py
@@ -21,8 +21,9 @@ from abc import abstractmethod
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
-from airflow.template.templater import Templater
+from airflow.sdk.definitions.templater import Templater
 from airflow.utils.context import context_merge
+from airflow.utils.log.logging_mixin import LoggingMixin
 
 if TYPE_CHECKING:
     import jinja2
@@ -31,7 +32,7 @@ if TYPE_CHECKING:
     from airflow.utils.context import Context
 
 
-class BaseNotifier(Templater):
+class BaseNotifier(LoggingMixin, Templater):
     """BaseNotifier class for sending notifications."""
 
     template_fields: Sequence[str] = ()
