+++ airflow/utils/mixins.py
@@ -20,13 +20,8 @@ from __future__ import annotations
 
 import multiprocessing
 import multiprocessing.context
-import typing
 
 from airflow.configuration import conf
-
-if typing.TYPE_CHECKING:
-    from airflow.models.operator import Operator
-    from airflow.utils.context import Context
 
 
 class MultiprocessingStartMethodMixin:
@@ -49,25 +44,3 @@ class MultiprocessingStartMethodMixin:
     def _get_multiprocessing_context(self) -> multiprocessing.context.DefaultContext:
         mp_start_method = self._get_multiprocessing_start_method()
         return multiprocessing.get_context(mp_start_method)  # type: ignore
-
-
-class ResolveMixin:
-    """A runtime-resolved value."""
-
-    def iter_references(self) -> typing.Iterable[tuple[Operator, str]]:
-        """
-        Find underlying XCom references this contains.
-
-        This is used by the DAG parser to recursively find task dependencies.
-
-        :meta private:
-        """
-        raise NotImplementedError
-
-    def resolve(self, context: Context, *, include_xcom: bool = True) -> typing.Any:
-        """
-        Resolve this value for runtime.
-
-        :meta private:
-        """
-        raise NotImplementedError
