+++ task_sdk/tests/defintions/test_macros.py
@@ -1,0 +1,72 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+from __future__ import annotations
+
+import lazy_object_proxy
+import pytest
+
+from airflow.sdk.definitions import macros
+
+
+@pytest.mark.parametrize(
+    "ds, days, expected",
+    [
+        ("2015-01-01", 5, "2015-01-06"),
+        ("2015-01-02", 0, "2015-01-02"),
+        ("2015-01-06", -5, "2015-01-01"),
+        (lazy_object_proxy.Proxy(lambda: "2015-01-01"), 5, "2015-01-06"),
+        (lazy_object_proxy.Proxy(lambda: "2015-01-02"), 0, "2015-01-02"),
+        (lazy_object_proxy.Proxy(lambda: "2015-01-06"), -5, "2015-01-01"),
+    ],
+)
+def test_ds_add(ds, days, expected):
+    result = macros.ds_add(ds, days)
+    assert result == expected
+
+
+@pytest.mark.parametrize(
+    "ds, input_format, output_format, expected",
+    [
+        ("2015-01-02", "%Y-%m-%d", "%m-%d-%y", "01-02-15"),
+        ("2015-01-02", "%Y-%m-%d", "%Y-%m-%d", "2015-01-02"),
+        ("1/5/2015", "%m/%d/%Y", "%m-%d-%y", "01-05-15"),
+        ("1/5/2015", "%m/%d/%Y", "%Y-%m-%d", "2015-01-05"),
+        (lazy_object_proxy.Proxy(lambda: "2015-01-02"), "%Y-%m-%d", "%m-%d-%y", "01-02-15"),
+        (lazy_object_proxy.Proxy(lambda: "2015-01-02"), "%Y-%m-%d", "%Y-%m-%d", "2015-01-02"),
+        (lazy_object_proxy.Proxy(lambda: "1/5/2015"), "%m/%d/%Y", "%m-%d-%y", "01-05-15"),
+        (lazy_object_proxy.Proxy(lambda: "1/5/2015"), "%m/%d/%Y", "%Y-%m-%d", "2015-01-05"),
+    ],
+)
+def test_ds_format(ds, input_format, output_format, expected):
+    result = macros.ds_format(ds, input_format, output_format)
+    assert result == expected
+
+
+@pytest.mark.parametrize(
+    "input_value, expected",
+    [
+        ('{"field1":"value1", "field2":4, "field3":true}', {"field1": "value1", "field2": 4, "field3": True}),
+        (
+            '{"field1": [ 1, 2, 3, 4, 5 ], "field2" : {"mini1" : 1, "mini2" : "2"}}',
+            {"field1": [1, 2, 3, 4, 5], "field2": {"mini1": 1, "mini2": "2"}},
+        ),
+    ],
+)
+def test_json_loads(input_value, expected):
+    result = macros.json.loads(input_value)
+    assert result == expected
