+++ tests/macros/test_macros.py
@@ -26,40 +26,6 @@ from airflow.utils import timezone
 
 
 @pytest.mark.parametrize(
-    "ds, days, expected",
-    [
-        ("2015-01-01", 5, "2015-01-06"),
-        ("2015-01-02", 0, "2015-01-02"),
-        ("2015-01-06", -5, "2015-01-01"),
-        (lazy_object_proxy.Proxy(lambda: "2015-01-01"), 5, "2015-01-06"),
-        (lazy_object_proxy.Proxy(lambda: "2015-01-02"), 0, "2015-01-02"),
-        (lazy_object_proxy.Proxy(lambda: "2015-01-06"), -5, "2015-01-01"),
-    ],
-)
-def test_ds_add(ds, days, expected):
-    result = macros.ds_add(ds, days)
-    assert result == expected
-
-
-@pytest.mark.parametrize(
-    "ds, input_format, output_format, expected",
-    [
-        ("2015-01-02", "%Y-%m-%d", "%m-%d-%y", "01-02-15"),
-        ("2015-01-02", "%Y-%m-%d", "%Y-%m-%d", "2015-01-02"),
-        ("1/5/2015", "%m/%d/%Y", "%m-%d-%y", "01-05-15"),
-        ("1/5/2015", "%m/%d/%Y", "%Y-%m-%d", "2015-01-05"),
-        (lazy_object_proxy.Proxy(lambda: "2015-01-02"), "%Y-%m-%d", "%m-%d-%y", "01-02-15"),
-        (lazy_object_proxy.Proxy(lambda: "2015-01-02"), "%Y-%m-%d", "%Y-%m-%d", "2015-01-02"),
-        (lazy_object_proxy.Proxy(lambda: "1/5/2015"), "%m/%d/%Y", "%m-%d-%y", "01-05-15"),
-        (lazy_object_proxy.Proxy(lambda: "1/5/2015"), "%m/%d/%Y", "%Y-%m-%d", "2015-01-05"),
-    ],
-)
-def test_ds_format(ds, input_format, output_format, expected):
-    result = macros.ds_format(ds, input_format, output_format)
-    assert result == expected
-
-
-@pytest.mark.parametrize(
     "ds, input_format, output_format, locale, expected",
     [
         ("2015-01-02", "%Y-%m-%d", "MM-dd-yy", None, "01-02-15"),
@@ -108,21 +74,6 @@ def test_ds_format_locale(ds, input_format, output_format, locale, expected):
 )
 def test_datetime_diff_for_humans(dt, since, expected):
     result = macros.datetime_diff_for_humans(dt, since)
-    assert result == expected
-
-
-@pytest.mark.parametrize(
-    "input_value, expected",
-    [
-        ('{"field1":"value1", "field2":4, "field3":true}', {"field1": "value1", "field2": 4, "field3": True}),
-        (
-            '{"field1": [ 1, 2, 3, 4, 5 ], "field2" : {"mini1" : 1, "mini2" : "2"}}',
-            {"field1": [1, 2, 3, 4, 5], "field2": {"mini1": 1, "mini2": "2"}},
-        ),
-    ],
-)
-def test_json_loads(input_value, expected):
-    result = macros.json.loads(input_value)
     assert result == expected
