+++ tests/models/test_baseoperator.py
@@ -18,14 +18,10 @@
 from __future__ import annotations
 
 import copy
-import logging
-import uuid
 from collections import defaultdict
-from datetime import date, datetime
-from typing import NamedTuple
+from datetime import datetime
 from unittest import mock
 
-import jinja2
 import pytest
 
 from airflow.decorators import task as task_decorator
@@ -44,7 +40,6 @@ from airflow.models.trigger import TriggerFailureReason
 from airflow.providers.common.sql.operators import sql
 from airflow.utils.edgemodifier import Label
 from airflow.utils.task_group import TaskGroup
-from airflow.utils.template import literal
 from airflow.utils.trigger_rule import TriggerRule
 from airflow.utils.types import DagRunType
 
@@ -72,17 +67,6 @@ class ClassWithCustomAttributes:
         return not self.__eq__(other)
 
 
-# Objects with circular references (for testing purpose)
-object1 = ClassWithCustomAttributes(attr="{{ foo }}_1", template_fields=["ref"])
-object2 = ClassWithCustomAttributes(attr="{{ foo }}_2", ref=object1, template_fields=["ref"])
-setattr(object1, "ref", object2)
-
-
-class MockNamedTuple(NamedTuple):
-    var1: str
-    var2: str
-
-
 class TestBaseOperator:
     def test_trigger_rule_validation(self):
         from airflow.models.abstractoperator import DEFAULT_TRIGGER_RULE
@@ -106,211 +90,6 @@ class TestBaseOperator:
         BaseOperator(
             task_id="test_valid_trigger_rule", dag=non_fail_stop_dag, trigger_rule=TriggerRule.ALWAYS
         )
-
-    @pytest.mark.db_test
-    @pytest.mark.parametrize(
-        ("content", "context", "expected_output"),
-        [
-            ("{{ foo }}", {"foo": "bar"}, "bar"),
-            (["{{ foo }}_1", "{{ foo }}_2"], {"foo": "bar"}, ["bar_1", "bar_2"]),
-            (("{{ foo }}_1", "{{ foo }}_2"), {"foo": "bar"}, ("bar_1", "bar_2")),
-            (
-                {"key1": "{{ foo }}_1", "key2": "{{ foo }}_2"},
-                {"foo": "bar"},
-                {"key1": "bar_1", "key2": "bar_2"},
-            ),
-            (
-                {"key_{{ foo }}_1": 1, "key_2": "{{ foo }}_2"},
-                {"foo": "bar"},
-                {"key_{{ foo }}_1": 1, "key_2": "bar_2"},
-            ),
-            (date(2018, 12, 6), {"foo": "bar"}, date(2018, 12, 6)),
-            (datetime(2018, 12, 6, 10, 55), {"foo": "bar"}, datetime(2018, 12, 6, 10, 55)),
-            (MockNamedTuple("{{ foo }}_1", "{{ foo }}_2"), {"foo": "bar"}, MockNamedTuple("bar_1", "bar_2")),
-            ({"{{ foo }}_1", "{{ foo }}_2"}, {"foo": "bar"}, {"bar_1", "bar_2"}),
-            (None, {}, None),
-            ([], {}, []),
-            ({}, {}, {}),
-            (
-                # check nested fields can be templated
-                ClassWithCustomAttributes(att1="{{ foo }}_1", att2="{{ foo }}_2", template_fields=["att1"]),
-                {"foo": "bar"},
-                ClassWithCustomAttributes(att1="bar_1", att2="{{ foo }}_2", template_fields=["att1"]),
-            ),
-            (
-                # check deep nested fields can be templated
-                ClassWithCustomAttributes(
-                    nested1=ClassWithCustomAttributes(
-                        att1="{{ foo }}_1", att2="{{ foo }}_2", template_fields=["att1"]
-                    ),
-                    nested2=ClassWithCustomAttributes(
-                        att3="{{ foo }}_3", att4="{{ foo }}_4", template_fields=["att3"]
-                    ),
-                    template_fields=["nested1"],
-                ),
-                {"foo": "bar"},
-                ClassWithCustomAttributes(
-                    nested1=ClassWithCustomAttributes(
-                        att1="bar_1", att2="{{ foo }}_2", template_fields=["att1"]
-                    ),
-                    nested2=ClassWithCustomAttributes(
-                        att3="{{ foo }}_3", att4="{{ foo }}_4", template_fields=["att3"]
-                    ),
-                    template_fields=["nested1"],
-                ),
-            ),
-            (
-                # check null value on nested template field
-                ClassWithCustomAttributes(att1=None, template_fields=["att1"]),
-                {},
-                ClassWithCustomAttributes(att1=None, template_fields=["att1"]),
-            ),
-            (
-                # check there is no RecursionError on circular references
-                object1,
-                {"foo": "bar"},
-                object1,
-            ),
-            # By default, Jinja2 drops one (single) trailing newline
-            ("{{ foo }}\n\n", {"foo": "bar"}, "bar\n"),
-            (literal("{{ foo }}"), {"foo": "bar"}, "{{ foo }}"),
-            (literal(["{{ foo }}_1", "{{ foo }}_2"]), {"foo": "bar"}, ["{{ foo }}_1", "{{ foo }}_2"]),
-            (literal(("{{ foo }}_1", "{{ foo }}_2")), {"foo": "bar"}, ("{{ foo }}_1", "{{ foo }}_2")),
-        ],
-    )
-    def test_render_template(self, content, context, expected_output):
-        """Test render_template given various input types."""
-        task = BaseOperator(task_id="op1")
-
-        result = task.render_template(content, context)
-        assert result == expected_output
-
-    @pytest.mark.parametrize(
-        ("content", "context", "expected_output"),
-        [
-            ("{{ foo }}", {"foo": "bar"}, "bar"),
-            ("{{ foo }}", {"foo": ["bar1", "bar2"]}, ["bar1", "bar2"]),
-            (["{{ foo }}", "{{ foo | length}}"], {"foo": ["bar1", "bar2"]}, [["bar1", "bar2"], 2]),
-            (("{{ foo }}_1", "{{ foo }}_2"), {"foo": "bar"}, ("bar_1", "bar_2")),
-            ("{{ ds }}", {"ds": date(2018, 12, 6)}, date(2018, 12, 6)),
-            (datetime(2018, 12, 6, 10, 55), {"foo": "bar"}, datetime(2018, 12, 6, 10, 55)),
-            ("{{ ds }}", {"ds": datetime(2018, 12, 6, 10, 55)}, datetime(2018, 12, 6, 10, 55)),
-            (MockNamedTuple("{{ foo }}_1", "{{ foo }}_2"), {"foo": "bar"}, MockNamedTuple("bar_1", "bar_2")),
-            (
-                ("{{ foo }}", "{{ foo.isoformat() }}"),
-                {"foo": datetime(2018, 12, 6, 10, 55)},
-                (datetime(2018, 12, 6, 10, 55), "2018-12-06T10:55:00"),
-            ),
-            (None, {}, None),
-            ([], {}, []),
-            ({}, {}, {}),
-        ],
-    )
-    def test_render_template_with_native_envs(self, content, context, expected_output):
-        """Test render_template given various input types with Native Python types"""
-        with DAG("test-dag", schedule=None, start_date=DEFAULT_DATE, render_template_as_native_obj=True):
-            task = BaseOperator(task_id="op1")
-
-        result = task.render_template(content, context)
-        assert result == expected_output
-
-    @pytest.mark.db_test
-    def test_render_template_fields(self):
-        """Verify if operator attributes are correctly templated."""
-        task = MockOperator(task_id="op1", arg1="{{ foo }}", arg2="{{ bar }}")
-
-        # Assert nothing is templated yet
-        assert task.arg1 == "{{ foo }}"
-        assert task.arg2 == "{{ bar }}"
-
-        # Trigger templating and verify if attributes are templated correctly
-        task.render_template_fields(context={"foo": "footemplated", "bar": "bartemplated"})
-        assert task.arg1 == "footemplated"
-        assert task.arg2 == "bartemplated"
-
-    @pytest.mark.db_test
-    def test_render_template_fields_func_using_context(self):
-        """Verify if operator attributes are correctly templated."""
-
-        def fn_to_template(context, jinja_env):
-            tmp = context["task"].render_template("{{ bar }}", context, jinja_env)
-            return "foo_" + tmp
-
-        task = MockOperator(task_id="op1", arg2=fn_to_template)
-
-        # Trigger templating and verify if attributes are templated correctly
-        task.render_template_fields(context={"bar": "bartemplated", "task": task})
-        assert task.arg2 == "foo_bartemplated"
-
-    @pytest.mark.db_test
-    def test_render_template_fields_simple_func(self):
-        """Verify if operator attributes are correctly templated."""
-
-        def fn_to_template(**kwargs):
-            a = "foo_" + ("bar" * 3)
-            return a
-
-        task = MockOperator(task_id="op1", arg2=fn_to_template)
-        task.render_template_fields({})
-        assert task.arg2 == "foo_barbarbar"
-
-    @pytest.mark.parametrize(("content",), [(object(),), (uuid.uuid4(),)])
-    def test_render_template_fields_no_change(self, content):
-        """Tests if non-templatable types remain unchanged."""
-        task = BaseOperator(task_id="op1")
-
-        result = task.render_template(content, {"foo": "bar"})
-        assert content is result
-
-    @pytest.mark.db_test
-    def test_nested_template_fields_declared_must_exist(self):
-        """Test render_template when a nested template field is missing."""
-        task = BaseOperator(task_id="op1")
-
-        error_message = (
-            "'missing_field' is configured as a template field but ClassWithCustomAttributes does not have "
-            "this attribute."
-        )
-        with pytest.raises(AttributeError, match=error_message):
-            task.render_template(
-                ClassWithCustomAttributes(
-                    template_fields=["missing_field"], task_type="ClassWithCustomAttributes"
-                ),
-                {},
-            )
-
-    def test_string_template_field_attr_is_converted_to_list(self):
-        """Verify template_fields attribute is converted to a list if declared as a string."""
-
-        class StringTemplateFieldsOperator(BaseOperator):
-            template_fields = "a_string"
-
-        warning_message = (
-            "The `template_fields` value for StringTemplateFieldsOperator is a string but should be a "
-            "list or tuple of string. Wrapping it in a list for execution. Please update "
-            "StringTemplateFieldsOperator accordingly."
-        )
-        with pytest.warns(UserWarning, match=warning_message) as warnings:
-            task = StringTemplateFieldsOperator(task_id="op1")
-
-            assert len(warnings) == 1
-            assert isinstance(task.template_fields, list)
-
-    def test_jinja_invalid_expression_is_just_propagated(self):
-        """Test render_template propagates Jinja invalid expression errors."""
-        task = BaseOperator(task_id="op1")
-
-        with pytest.raises(jinja2.exceptions.TemplateSyntaxError):
-            task.render_template("{{ invalid expression }}", {})
-
-    @pytest.mark.db_test
-    @mock.patch("airflow.templates.SandboxedEnvironment", autospec=True)
-    def test_jinja_env_creation(self, mock_jinja_env):
-        """Verify if a Jinja environment is created only once when templating."""
-        task = MockOperator(task_id="op1", arg1="{{ foo }}", arg2="{{ bar }}")
-
-        task.render_template_fields(context={"foo": "whatever", "bar": "whatever"})
-        assert mock_jinja_env.call_count == 1
 
     def test_cross_downstream(self):
         """Test if all dependencies between tasks are all set correctly."""
@@ -630,63 +409,6 @@ def test_deepcopy():
 
         MockOperator(task_id="task1", arg1=task0())
     copy.deepcopy(dag)
-
-
-@pytest.mark.db_test
-@pytest.mark.parametrize(
-    ("task", "context", "expected_exception", "expected_rendering", "expected_log", "not_expected_log"),
-    [
-        # Simple success case.
-        (
-            MockOperator(task_id="op1", arg1="{{ foo }}"),
-            dict(foo="footemplated"),
-            None,
-            dict(arg1="footemplated"),
-            None,
-            "Exception rendering Jinja template",
-        ),
-        # Jinja syntax error.
-        (
-            MockOperator(task_id="op1", arg1="{{ foo"),
-            dict(),
-            jinja2.TemplateSyntaxError,
-            None,
-            "Exception rendering Jinja template for task 'op1', field 'arg1'. Template: '{{ foo'",
-            None,
-        ),
-        # Type error
-        (
-            MockOperator(task_id="op1", arg1="{{ foo + 1 }}"),
-            dict(foo="footemplated"),
-            TypeError,
-            None,
-            "Exception rendering Jinja template for task 'op1', field 'arg1'. Template: '{{ foo + 1 }}'",
-            None,
-        ),
-    ],
-)
-def test_render_template_fields_logging(
-    caplog, monkeypatch, task, context, expected_exception, expected_rendering, expected_log, not_expected_log
-):
-    """Verify if operator attributes are correctly templated."""
-
-    # Trigger templating and verify results
-    def _do_render():
-        task.render_template_fields(context=context)
-
-    logger = logging.getLogger("airflow.task")
-    monkeypatch.setattr(logger, "propagate", True)
-    if expected_exception:
-        with pytest.raises(expected_exception):
-            _do_render()
-    else:
-        _do_render()
-        for k, v in expected_rendering.items():
-            assert getattr(task, k) == v
-    if expected_log:
-        assert expected_log in caplog.text
-    if not_expected_log:
-        assert not_expected_log not in caplog.text
 
 
 @pytest.mark.db_test
